+++ src/util.cc
@@ -1887,7 +1887,7 @@ uint8_t msghdr_get_ecn(msghdr *msg, int family) {
 }
 
 size_t msghdr_get_udp_gro(msghdr *msg) {
-  uint16_t gso_size = 0;
+  int gso_size = 0;
 
 #  ifdef UDP_GRO
   for (auto cmsg = CMSG_FIRSTHDR(msg); cmsg; cmsg = CMSG_NXTHDR(msg, cmsg)) {
@@ -1899,7 +1899,7 @@ size_t msghdr_get_udp_gro(msghdr *msg) {
   }
 #  endif // UDP_GRO
 
-  return gso_size;
+  return static_cast<size_t>(gso_size);
 }
 #endif // ENABLE_HTTP3
